Revert "Revert "Merge pull request #99 from mlfoundations/interleaved_dataloader_pr1" (#105)"
This reverts commit 1ce3e8c7657995b271ad775bcc2a33b95dffa28d.
